Refactor code, rename package.
